requirements.txt
python-telegram-bot==20.x
requests==2.31.0
